% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyRescale.R
\name{FuzzyRescale}
\alias{FuzzyRescale}
\title{R6 Fuzzy Rescale}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for
vectors, matrices, and data frames. Factors are allowed, but
characters are not.}

\item{select}{expression, indicating columns to select from a data frame or
matrix}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. This class is
missing tolerable (i.e., compatible), in which the
transformation will not remove \code{NAs} if na.rm is FALSE, it
abides the dimension of vectors and retrun NAs beside the truly
transformed values.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{max}{a scalar indicating the maximum endpoint of rescaled variable. The
might be 1, 100, etc.}

\item{center}{it determines the central value of a fuzzy set. It should be
any of the values \link[base:mean]{"mean"},
\link[stats:median]{"median"} or any numeric input. It will have
a membership degree of 1, in \code{gaussmf()} method; 0.5 in
\code{sigmf()};}

\item{sigma}{it determines the width of a fuzzy set obtained from either
\code{gaussmf()} or \code{sigmf()}. It should be any of the
values \link[stats:sd]{"sd"}, \link[stats:IQR]{"IQR"} or any
numeric input. In \code{sigmf()}, to open the membership function
to the left or right, specify a negative or positive value for
\code{sigma}, respectively.}

\item{param_one}{This is the first parameter for \code{smf()}. If
\code{param_one} is lower than \code{param_two}, it will be the
\strong{foot} of the membership function; otherwise it will define
the \strong{shoulder}. It should be either \link[base:min]{"min"}
or \link[base:max]{"max"}, any
\link[cvcqv:SampleQuantiles]{quantiles} defined by the prefix
\code{"p"} followed by a value in \code{[0,100]} (\emph{e.g.,}
"p2.5" indicates 2.5\% percentile of the input variable),
or any numeric input.}

\item{param_two}{This is the second parameter for \code{smf()}. If
\code{param_two} is higher than \code{param_one}, it will define
the \strong{shoulder} of the membership function; otherwise it
will be the \strong{foot} It should be either
\link[base:max]{"max"} or \link[base:min]{"min"}, any
\link[cvcqv:SampleQuantiles]{quantiles} defined by the prefix
\code{"p"} followed by a value in \code{[0,100]} (\emph{e.g.,}
"p97.5" indicates 97.5\% percentile of the input variable),
or any numeric input.}
}
\description{
The R6 class \code{FuzzyRescale} transforms data into 0-1 or
             0-100 scale based on Fuzzy membership functions.
}
\examples{
x <- c(
   0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
   4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
x_fuzzy <- FuzzyRescale$new(x)
R6::is.R6(x_fuzzy)
x_fuzzy$gaussmf()
}
