% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileRescale.R
\name{QuantileRescale}
\alias{QuantileRescale}
\title{R6 Quantile Rescale}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for numeric
vectors, matrices, and data frames.}

\item{select}{expression, indicating columns to select from a data frame or
matrix}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. This class is
missing tolerable (i.e., compatible), in which the
transformation will not remove \code{NAs} if na.rm is FALSE, it
abides the dimension of vectors and retrun NAs beside the truly
transformed values.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{probs}{numeric vector of probabilities with values in \code{[0,1]}.}

\item{max}{a scalar indicating the maximum endpoint of rescaled variable. The
might be 1, 100, etc.}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile
algorithms explained in \link[stats]{quantile} to be used.}
}
\description{
The R6 class \code{QuantileRescale} transforms data into 0-1 or
             0-100 scale with values above (1-p)\% and below p\% percentiles
             taken as the nearest endpoints.
}
\examples{
x <- c(
   0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
   4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
x_trf <- QuantileRescale$new(x)
R6::is.R6(x_trf)
x_trf$transform_x()
}
