% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyRescale.R
\name{FuzzyRescale}
\alias{FuzzyRescale}
\title{R6 Fuzzy Rescale}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for
vectors, matrices, and data frames. Factors are allowed, but
characters are not.}

\item{select}{expression, indicating columns to select from a data frame or
matrix}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. This class is
missing tolerable (i.e., compatible), in which the
transformation will not remove \code{NAs} if na.rm is FALSE, it
abides the dimension of vectors and retrun NAs beside the truly
transformed values.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{max}{a scalar indicating the maximum endpoint of rescaled variable. The
might be 1, 100, etc.}

\item{center}{it determines the central value of a fuzzy set with
\code{gaussmf()} method, and will have a membership degree of 1.
It should be any of the values "mean", "median" or any numeric
input.}

\item{sigma}{it determines the width of a fuzzy set with
\code{gaussmf()} method. It should be any of the values
"sd", "IQR" or any numeric input.}
}
\description{
The R6 class \code{FuzzyRescale} transforms data into 0-1 or
             0-100 scale based on Fuzzy membership functions.
}
\examples{
x <- c(
   0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
   4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
x_fuzzy <- FuzzyRescale$new(x)
R6::is.R6(x_fuzzy)
x_fuzzy$gaussmf()
}
